FEATURE-Refactoring: Refactoring tkinter application to Webapp using flask
