+++ README.md
@@ -1,26 +1,55 @@
 # ace-est-fitness-and-gym
 An automated application that demonstrates core functionalities pertinent to a fitness and gym management system.
 
-# Install pyenv
+## Python installation & local setup
+### Install pyenv
+```
 brew install pyenv
+```
 
-# Add these lines (in this exact order) in ~/.zshrc:
+### Add these lines (in this exact order) in ~/.zshrc:
+```
 export PYENV_ROOT="$HOME/.pyenv"
 export PATH="$PYENV_ROOT/bin:$PATH"
 eval "$(pyenv init -)"
 eval "$(pyenv init --path)"
+```
 
-# Reload shell
+### Reload shell
+```
 source ~/.zshrc
+```
 
-# Install Python
+### Install Python
+```
 pyenv install 3.13.0
 pyenv global 3.13.0
+```
 
-# Setting Vitual Environment for python runtime
+### Setting Vitual Environment for python runtime
+```
 mkdir .env
 python3 -m venv .env
 source .env/bin/activate
+```
 
-# Intall required python libraries
+### Intall required python libraries
+```
 pip install -r requirements.txt
+```
+
+## Running & testing the application locally
+### To run the application locally
+```
+python3 src/app.py
+```
+
+### To run unit test cases locally
+```
+.env/bin/python -m pytest tests/test_workout_manager.py
+```
+
+### To clean up runtime files generated by pytest 
+```
+rm -rf __pycache__ */__pycache__ .pytest_cache */.pytest_cache
+```
