--- commands.txt
@@ -1,25 +1,0 @@
-# To run the application
-python3 src/ACEest_Fitness.py 
-
-# To run unit test cases
-".env/bin/python" -m pytest tests/test_workout_manager.py
-
-# To clean up runtime files generated by pytest 
-rm -rf __pycache__ */__pycache__ .pytest_cache
-
-# GIT Commands
-
-# To see the status of files
-git status
-
-# To stage all changes, including new files, modifications, and deletions
-git add -A
-
-# To stage all modifications, including deletions, for tracked files and its subdirectories. It does not stage new, untracked files.
-git add -u
-
-# To stage a specific file
-git add file_name
-
-# To commit the changes
-git commit -m "commit message"
