+++ src/templates/index.html
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>ACEestFitness and Gym</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
+</head>
+<body class="bg-light">
+    <div class="container mt-5">
+        <h1 class="mb-4">ACEestFitness and Gym</h1>
+        {% with messages = get_flashed_messages(with_categories=true) %}
+          {% if messages %}
+            {% for category, message in messages %}
+              <div class="alert alert-{{ category }}">{{ message }}</div>
+            {% endfor %}
+          {% endif %}
+        {% endwith %}
+        <form method="post" class="mb-4">
+            <div class="mb-3">
+                <label for="workout" class="form-label">Workout</label>
+                <input type="text" class="form-control" id="workout" name="workout" required>
+            </div>
+            <div class="mb-3">
+                <label for="duration" class="form-label">Duration (minutes)</label>
+                <input type="number" class="form-control" id="duration" name="duration" required>
+            </div>
+            <button type="submit" class="btn btn-primary">Add Workout</button>
+        </form>
+        <h2>Logged Workouts</h2>
+        {% if workouts %}
+            <ul class="list-group">
+            {% for entry in workouts %}
+                <li class="list-group-item">{{ entry.workout }} - {{ entry.duration }} minutes</li>
+            {% endfor %}
+            </ul>
+        {% else %}
+            <p>No workouts logged yet.</p>
+        {% endif %}
+    </div>
+</body>
+</html>
